topology: remove it
