+++ client.go
@@ -2,6 +2,7 @@ package egoscale
 
 import (
 	"context"
+	"crypto/tls"
 	"net/http"
 	"time"
 )
@@ -54,4 +55,52 @@ func (client *Client) Delete(g Deletable) error {
 // DeleteWithContext removes the given resource of fails
 func (client *Client) DeleteWithContext(ctx context.Context, g Deletable) error {
 	return g.Delete(ctx, client)
+}
+
+// RetryStrategyFunc represents a how much time to wait between two calls to CloudStack
+type RetryStrategyFunc func(int64) time.Duration
+
+// NewClientWithTimeout creates a CloudStack API client
+//
+// Timeout is set to booth the HTTP client and the client itself.
+func NewClientWithTimeout(endpoint, apiKey, apiSecret string, timeout time.Duration) *Client {
+	client := &http.Client{
+		Timeout: timeout,
+		Transport: &http.Transport{
+			Proxy: http.ProxyFromEnvironment,
+			TLSClientConfig: &tls.Config{
+				InsecureSkipVerify: false,
+			},
+		},
+	}
+
+	cs := &Client{
+		client:        client,
+		endpoint:      endpoint,
+		apiKey:        apiKey,
+		apiSecret:     apiSecret,
+		Timeout:       timeout,
+		RetryStrategy: FibonacciRetryStrategy,
+	}
+
+	return cs
+}
+
+// NewClient creates a CloudStack API client with default timeout (60)
+func NewClient(endpoint, apiKey, apiSecret string) *Client {
+	timeout := time.Duration(60 * time.Second)
+	return NewClientWithTimeout(endpoint, apiKey, apiSecret, timeout)
+}
+
+// FibonacciRetryStrategy waits for an increasing amount of time following the Fibonacci sequence
+func FibonacciRetryStrategy(iteration int64) time.Duration {
+	var a, b, i, tmp int64
+	a = 0
+	b = 1
+	for i = 0; i < iteration; i++ {
+		tmp = a + b
+		a = b
+		b = tmp
+	}
+	return time.Duration(a) * time.Second
 }
